MVC architecture + route handling
