+++ controllers/userController.js
@@ -1,0 +1,5 @@
+const User = require('../models/User');
+
+exports.login = (req, res) => {
+	res.render('login');
+};
