+++ models/User.js
@@ -1,0 +1,15 @@
+const mongoose = require('mongoose');
+const passportLocalMongoose = require('passport-local-mongoose');
+const findOrCreate = require("mongoose-findorcreate");
+
+const UserSchema = new mongoose.Schema({
+	fullname: String,
+	email: String,
+	password: String,
+	role: String
+});
+
+UserSchema.plugin(passportLocalMongoose);
+UserSchema.plugin(findOrCreate);
+
+module.exports = mongoose.model('User', UserSchema);
