--- public/css/signup.css
@@ -1,73 +1,0 @@
-body {
-    background: linear-gradient(230.3deg, #01265B 10.13%, #0251C1 88.77%);
-    padding: 80px 70px 80px 70px;
-    display:flex;
-    justify-content: center;
-    align-items: center;
-    flex-direction: row;
-}
-.body--icon {
-    align-items: flex-start;
-    width: 350px;
-    height: 400px;
-}
-
-.body--form{
-    display:flex;
-    flex-direction: column;
-    align-items: center;
-    border-radius: 20px;
-    width: 550px;
-    height: 400px;
-    background: #F0F3F6;
-    box-shadow: 20px 30px 40px 0px #00000040;
-}
-
-.body--form input[type="text"],
-.body--form input[type="password"]{
-
-    /*margin-right: 150px;*/
-    width: 350px;
-    height: 25px;
-    border-radius: 10px;
-    border: none;
-    font-family: Lexend;
-    font-size: 16px;
-    font-weight: 400;
-    line-height: 20px;
-    text-align: center;
-}
-
-.body--button {
-    background: linear-gradient(90deg, #01265B 0%, #0251C1 100%);
-    width: 250px;
-    height: 30px;
-    top: 529px;
-    left: 693px;
-    text-align: center;
-    border-radius: 7px;
-    font-family: Lexend;
-    font-size: 16px;
-    font-weight: 600;
-    line-height: 20px;
-    color: #FFFFFF;
-}
-
-.body--ending {
-    font-family: Lexend;
-    font-size: 12px;
-    font-weight: 600;
-    line-height: 15px;
-    text-align: left;
-    color: #AAAAAA;
-}
-
-.body--login{
-    margin-top: 20px;
-    margin-bottom: 10px;
-    font-family: Lexend Deca;
-    font-size: 40px;
-    font-weight: 500;
-    line-height: 50px;
-    color: #000000;
-}
