+++ routes/adminRoutes.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+const adminController = require('../controllers/adminController');
+
+router.get('/', adminController.adminDashboard);
+router.post('/create-user', adminController.createUser);
+
+module.exports = router;
