+++ views/admin.ejs
@@ -1,4 +1,4 @@
-<form action="/admin" method="post" class="rectangle">
+<form action="/admin/create-user" method="post" class="rectangle">
     <p>
         <label for="userInput"></label>
         <input type="text" id="userInput" name="username" placeholder="Phone number or email" required >
