--- views/signup.ejs
@@ -1,61 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Sign Up</title>
-    <link rel="stylesheet" href="../public/css/signup.css" type="text/css">
-    <script>
-        function validateInput() {
-            var userInput = document.getElementById("show").value;
-            var emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
-            var phoneRegex = /^\+355\d{8}$/; // Phone number starts with +355 and has 8 digits after that
-
-            // Check if the input matches the email regex or phone number regex
-            if (emailRegex.test(userInput) || phoneRegex.test(userInput)) {
-                alert("Input is valid!");
-            } else {
-                alert("Please enter a valid email address or phone number.");
-            }
-        }
-    </script>
-    <script>
-        function myfunction(){
-            var show = document.getElementById('userInput');
-            if (show.type== 'password'){
-                show.type='text';
-            }
-            else{
-                show.type='password';
-            }
-        }
-    </script>
-</head>
-<body >
-<img src="../public/css/firstphoto.png" class="body--icon"/>
-
-<form action="/signup" method="post" onsubmit="return validateInput();" class="body--form"/>
-<p class="body--login">Sign up</p>
-    <p>
-        <label for="name"></label>
-        <input type="text" name="fullname" id="name" placeholder="Full name" autocomplete="on" required>
-
-    </p>
-    <p><label for="userInput"></label>
-        <input type="text" id="show" name="username" placeholder="Phone number or email" required >
-    </p>
-    <p>
-        <label for=""></label>
-        <input type="password" id="userInput" name="password" required placeholder="Password"> <br>
-        <input type="checkbox" name="" onclick="myfunction()" id="">
-        <label for="">Show Password</label>
-    </p>
-    <p>
-        <input class="body--button" type="submit" value="Sign up">
-    </p>
-<p class="body--ending ">Already have an account?</p>
-<a href="login.ejs"> Log in</a>
-
-</form>
-
-</body>
-</html>
